+++ app/src/pages/inventory/SoftwareInvPage.jsx
@@ -64,7 +64,6 @@ function SoftwareInvPage() {
     const formData = new FormData(e.target);
     const data = Object.fromEntries(formData.entries());
 
-    console.log("data modified", data)
     const updatedItem = {
       name: data.name,
       version: data.version,
@@ -88,11 +87,15 @@ function SoftwareInvPage() {
   }
 
   async function handleRemove(id) {
-    const deleted = await softwareApi.deleteSoftware(id);
-    if (!deleted) return;
-    setSoftware(prev => prev.filter(el => el._id !== id))
+    const userConfirmation = confirm(`¿Seguro que quieres proceder a eliminar el software cuya id es ${id}?`);
+    if (userConfirmation) {
+      const deleted = await softwareApi.deleteSoftware(id);
+      if (!deleted) return;
+      setSoftware(prev => prev.filter(el => el._id !== id))
 
-    await syncRemoveWithHardwareAndServers(id);
+      await syncRemoveWithHardwareAndServers(id);
+    }
+
   }
 
   return (
